Se devolvio la clase IngSalEnfermeras
Ademas se finalizo con la implementacion de este, haciendo que calcule las horas trabajadas diarias.

+++ src/main/java/Proyecto_progra/Contrato.java
@@ -11,7 +11,7 @@ package Proyecto_progra;
  */
 public class Contrato extends Salario {
     
-    public int calcularSueldo( int horas){
+    public int calcularSueldo( double horas){
     
         return 1;
     }
